Speed and Heading Control
Controller.py has the control algorithm which implements a PID controller for the Speed Control of the Autonomous Car and Stanley Controller that Controls the steering angle for the Car. Racetrack_waypoints is the input coordinates and velocity profile and trajectory is the output generated at the end of the episode.